Refactor into separate rust module files for exercise-contract
