+++ sdk/daml-lf/interpreter/src/test/rust/README.md
@@ -32,7 +32,6 @@ Note:
 In the Rust package directory `./hello-world` run:
 ```shell
 mkdir -p ./src/lf
-protoc --rust_out ./src/lf --proto_path ../../../../../transaction/src/main/protobuf/com/digitalasset/daml/lf value.proto
 cargo build --target wasm32-unknown-unknown --release
 cp ./target/wasm32-unknown-unknown/release/hello_world.wasm ../../resources/hello-world.rs.wasm
 ```
@@ -42,7 +41,7 @@ cp ./target/wasm32-unknown-unknown/release/hello_world.wasm ../../resources/hell
 In the Rust package directory `./create-contract` run:
 ```shell
 mkdir -p ./src/lf
-protoc --rust_out ./src/lf --proto_path ../../../../../transaction/src/main/protobuf/com/digitalasset/daml/lf value.proto
+protoc --rust_out ./src/lf --proto_path ../../../../../transaction/src/main/protobuf/com/digitalasset/daml/protobuf value.proto
 cargo build --target wasm32-unknown-unknown --release
 cp ./target/wasm32-unknown-unknown/release/create_contract.wasm ../../resources/create-contract.rs.wasm
 ```
@@ -52,7 +51,7 @@ cp ./target/wasm32-unknown-unknown/release/create_contract.wasm ../../resources/
 In the Rust package directory `./fetch-contract` run:
 ```shell
 mkdir -p ./src/lf
-protoc --rust_out ./src/lf --proto_path ../../../../../transaction/src/main/protobuf/com/digitalasset/daml/lf value.proto
+protoc --rust_out ./src/lf --proto_path ../../../../../transaction/src/main/protobuf/com/digitalasset/daml/protobuf value.proto
 cargo build --target wasm32-unknown-unknown --release
 cp ./target/wasm32-unknown-unknown/release/fetch_contract.wasm ../../resources/fetch-contract.rs.wasm
 ```
@@ -62,7 +61,7 @@ cp ./target/wasm32-unknown-unknown/release/fetch_contract.wasm ../../resources/f
 In the Rust package directory `./exercise-choice` run:
 ```shell
 mkdir -p ./src/lf
-protoc --rust_out ./src/lf --proto_path ../../../../../transaction/src/main/protobuf/com/digitalasset/daml/lf value.proto
+protoc --rust_out ./src/lf --proto_path ../../../../../transaction/src/main/protobuf/com/digitalasset/daml/protobuf value.proto
 cargo build --target wasm32-unknown-unknown --release
 cp ./target/wasm32-unknown-unknown/release/exercise_choice.wasm ../../resources/exercise-choice.rs.wasm
 ```
