--- sdk/daml-lf/interpreter/src/test/rust/exercise-choice/exercise-choice/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "exercise-choice"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
