+++ sdk/daml-lf/interpreter/src/test/rust/exercise-choice/src/protobuf/value.rs
